Moved testing of control specification into the driver, as to remove circular dependency
