+++ specification/k3po.control/pom.xml
@@ -20,12 +20,6 @@
     </scm>
 
     <dependencies>
-        <dependency>
-            <groupId>org.kaazing</groupId>
-            <artifactId>k3po.junit</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <build>
@@ -35,22 +29,6 @@
             </resource>
         </resources>
         <plugins>
-            <plugin>
-                <groupId>org.kaazing</groupId>
-                <artifactId>k3po-maven-plugin</artifactId>
-                <version>${project.version}</version>
-                <configuration>
-                    <skipTests>${skipITs}</skipTests>
-                </configuration>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>start</goal>
-                            <goal>stop</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
             <plugin>
                 <artifactId>maven-failsafe-plugin</artifactId>
                 <configuration>
